+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - london marathon bot: parse event id
 - async details fetcher can take note of urls too
+- get_time_eta as hh:mm:ss format
 
 ## 0.0 - 2017-04-25